hdfs,main: add 'hadmin fhs'
'hadmin fhs' checks for problems with standard directories and
permissions in HDFS, and optionally fixes them up.

For example, there should be a /tmp that is owned by 'hdfs' and has the
sticky bit set; 'hadmin fhs' will warn about problems with this and fix
the problem if '--fixup' is passed to the command.